TST: Divide tests by widget. Add striptool value reset test
